Initial version
This image is supposed to serve as a base for Jenkins nodes with
specific capabilities, such as being able to run a certain program or
compile executables with gcc.

To start the node, the environment variables NODE_NAME and NODE_SECRET
have to be set accordingly. The environment variable JENKINS_URL can be
overridden if the node is supposed to serve a Jenkins master different
from ImageJ's.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>